+++ src/app/modules/accounts/components/account/account.component.ts
@@ -1,12 +1,10 @@
-import { AfterViewInit, Component, OnInit } from '@angular/core';
+import { AfterViewInit, Component, Inject, OnInit } from '@angular/core';
 import { FormBuilder, FormControl, FormGroup, ReactiveFormsModule } from '@angular/forms';
 import { MatDatepickerModule } from '@angular/material/datepicker';
-import { MatDialogRef } from '@angular/material/dialog';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { MatIconModule } from '@angular/material/icon';
 import { MatInputModule } from '@angular/material/input';
 import { MatRadioModule } from '@angular/material/radio';
-import { IType } from '@shared/interfaces/type';
-import { CURRENCIES } from '../../../../local/constants/constants';
 import { MatSelectModule } from '@angular/material/select';
 import { MatDividerModule } from '@angular/material/divider';
 import { provideNativeDateAdapter } from '@angular/material/core';
@@ -14,6 +12,8 @@ import { MatButtonModule } from '@angular/material/button';
 import { IAccount } from '@modules/accounts/interfaces/accounts';
 import { AccountMapper } from '@modules/accounts/mappers/account.mapper';
 import { AccountsService } from '@modules/accounts/services/accounts.service';
+import { IAccountType, ICurrency } from '@modules/administration/interfaces/creation-type';
+import { AdministrationService } from '@modules/administration/services/administration.service';
 
 @Component({
   selector: 'app-account',
@@ -33,27 +33,27 @@ import { AccountsService } from '@modules/accounts/services/accounts.service';
 })
 export class AccountComponent implements OnInit, AfterViewInit {
 
-  // ******************************* PROVISIONAL PROPERTIES *******************************
-  accountTypes: IType[] = [ 
-    { value: 1, type: 'cash' },
-    { value: 2, type: 'bank' }
-  ];
-
-  public CURRENCIES = CURRENCIES
-
-  // ******************************* FORM RELATED PROPERTIES *******************************
+  // ******************************* PROPERTIES *******************************
   public accountForm!: FormGroup
+  public accountTypes: IAccountType[] = []
+  public currencies: ICurrency[] = []
+  public row!: IAccount
 
   constructor(
+    @Inject(MAT_DIALOG_DATA) public data: IAccount, 
     public dialogRef: MatDialogRef<AccountComponent>,
     private readonly fb: FormBuilder,
+    private readonly administrationService: AdministrationService,
     private readonly accountService: AccountsService
-  ) { }
+  ) {
+    this.row = data
+  }
 
   // ******************************* HOOKS *******************************
   ngOnInit(): void {
+    this.getAccountTypes()
+    this.getCurrencies()
     this.createAccountForm()
-    this.accountForm.valueChanges.subscribe(() => console.log(this.accountForm.getRawValue()))
   }
 
   ngAfterViewInit(): void {
@@ -62,14 +62,25 @@ export class AccountComponent implements OnInit, AfterViewInit {
 
   // ******************************* MAIN METHODS *******************************
   /**
+   * Fetches the existing account types
+   */
+  private async getAccountTypes() {
+    this.accountTypes = await this.administrationService.getAccountTypes()
+  }
+
+  private async getCurrencies() {
+    this.currencies = await this.administrationService.getCurrencies()
+  }
+
+  /**
    * Creates the fields for the account form
    */
   private createAccountForm() {
     this.accountForm = this.fb.group({
       created_at: new FormControl({ value: null, disabled: true }),
       name: new FormControl({ value: null, disabled: true }),
-      type: new FormControl({ value: null, disabled: true }),
-      currency: new FormControl({ value: null, disabled: true }),
+      type_id: new FormControl({ value: null, disabled: true }),
+      currency_code: new FormControl({ value: null, disabled: true }),
       initial_balance: new FormControl({ value: null, disabled: true }),
       description: new FormControl({ value: null, disabled: true }),
       active: new FormControl({ value: true, disabled: true })
@@ -81,16 +92,16 @@ export class AccountComponent implements OnInit, AfterViewInit {
    */
   public async saveAccount() {
     try {
-      const formValue = this.accountForm.value
-      const account: IAccount = AccountMapper.fromForm(formValue)
+      const formValue = this.accountForm.getRawValue()
+      const account: IAccount = AccountMapper.fromForm(formValue, this.row)
 
       await this.accountService.saveAccount(account)
       this.accountForm.reset()
-      this.close(true)
+      this.dialogRef.close({success: true})
     }
     catch (error) {
       console.warn("Error saving account", error)
-      this.close(false)
+      this.dialogRef.close({success: false})
     }
   }
 
@@ -98,8 +109,8 @@ export class AccountComponent implements OnInit, AfterViewInit {
    * Closes the modal
    * @param success Indicates whether the operation was successful or not
    */
-  public close(success: boolean) {
-    this.dialogRef.close(success)
+  public close() {
+    this.dialogRef.close()
   }
 
   // ******************************* HELPERS *******************************