+++ src/app/modules/accounts/interfaces/accounts.ts
@@ -1,20 +1,25 @@
 export interface IAccount {
-    id?: string
+    id?: number
     name: string
-    type: string
-    currency: string
+    type_id: number
+    currency_code: string
     initial_balance: number
     description: string
     active: boolean
-    created_at: string
+    created_at: string,
+    updated_at?: string
+}
+
+export interface IAccountDisplay extends IAccount {
+    type_name: string
 }
 
 export interface IAccountFormModel {
     created_at: Date
     name: string
-    type: string
-    currency: string
-    initial_balance: number
+    type_id: number
+    currency_code: string
+    initial_balance: string
     description: string
     active: boolean
 }